moved mapping to EasyToDataverseMapper class
